refactor(api): extract utils from test-helper.js
