+++ api/tests/tooling/test-utils/wait.js
@@ -1,0 +1,11 @@
+export function wait(ms) {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, ms);
+  });
+}
+
+export function waitForStreamFinalizationToBeDone() {
+  return wait(300);
+}
