+++ api/tests/unit/tooling/test-utils/error_test.js
@@ -1,6 +1,6 @@
-import { catchErr, expect, parseJsonStream } from './test-helper.js';
+import { catchErr, expect } from '../../../test-helper.js';
 
-describe('Test helpers', function () {
+describe('Error test utils', function () {
   describe('#catchErr', function () {
     it('returns the error thrown in the tested function', async function () {
       // given
@@ -27,20 +27,6 @@ describe('Test helpers', function () {
 
       // then
       await expect(promise).to.be.rejectedWith('Expected an error, but none was thrown.');
-    });
-  });
-
-  describe('#parseJsonStream', function () {
-    it('should parse JSONStream data', function () {
-      const obj1 = { a: 1 };
-      const obj2 = { b: 2 };
-      const data = [JSON.stringify(obj1), JSON.stringify(obj2), ''].join('\n');
-
-      expect(
-        parseJsonStream({
-          result: data,
-        }),
-      ).to.deep.equal([obj1, obj2]);
     });
   });
 });
